+++ cnccode/business_classes/BUCustomerReviewMeeting.inc.php
@@ -165,14 +165,22 @@ class BUCustomerReviewMeeting extends Business
             $hdrs = array(
                 'From' => $senderEmail,
                 'Subject' => $subject,
-                'Date' => date("r")
+                'Date' => date("r"),
+                'Content-Type' => 'text/html; charset=UTF-8'
             );
 
             $buMail->mime->setHTMLBody($body);
 
             $buMail->mime->addAttachment($icsFile, 'text/calendar', 'meeting.ics', false);
 
-            $body = $buMail->mime->get();
+            $mime_params = array(
+                'text_encoding' => '7bit',
+                'text_charset' => 'UTF-8',
+                'html_charset' => 'UTF-8',
+                'head_charset' => 'UTF-8'
+            );
+
+            $body = $buMail->mime->get($mime_params);
 
             $hdrs = $buMail->mime->headers($hdrs);