+++ cnccode/business_classes/BUDailyReport.inc.php
@@ -593,7 +593,8 @@ class BUDailyReport extends Business
         $hdrs = array(
             'From' => $senderEmail,
             'Subject' => $subject,
-            'Date' => date("r")
+            'Date' => date("r"),
+            'Content-Type' => 'text/html; charset=UTF-8'
         );
 
         $buMail->mime->setHTMLBody($body);
@@ -602,7 +603,13 @@ class BUDailyReport extends Business
             $buMail->mime->addAttachment($attachment, 'text/plain', 'report.csv', false);
         }
 
-        $body = $buMail->mime->get();
+        $mime_params = array(
+            'text_encoding' => '7bit',
+            'text_charset' => 'UTF-8',
+            'html_charset' => 'UTF-8',
+            'head_charset' => 'UTF-8'
+        );
+        $body = $buMail->mime->get($mime_params);
 
         $hdrs = $buMail->mime->headers($hdrs);