+++ cnccode/business_classes/BUMonthlyReport.inc.php
@@ -173,14 +173,21 @@ class BUMonthlyReport extends Business
             'From' => $senderEmail,
             'To' => $toEmail,
             'Subject' => $subject,
-            'Date' => date("r")
+            'Date' => date("r"),
+            'Content-Type' => 'text/html; charset=UTF-8'
         );
 
         $body = '<A HREF="' . $link . '">Monthly Report</A>';
 
         $buMail->mime->setHTMLBody($body);
 
-        $body = $buMail->mime->get();
+        $mime_params = array(
+            'text_encoding' => '7bit',
+            'text_charset' => 'UTF-8',
+            'html_charset' => 'UTF-8',
+            'head_charset' => 'UTF-8'
+        );
+        $body = $buMail->mime->get($mime_params);
 
         $hdrs = $buMail->mime->headers($hdrs);