+++ cnccode/business_classes/BUPDFSalesQuote.inc.php
@@ -488,14 +488,21 @@ class BUPDFSalesQuote extends Business
             'From' => $senderName . '<' . $senderEmail . '>',
             'To' => $toEmail,
             'Subject' => $subject,
-            'Date' => date("r")
+            'Date' => date("r"),
+            'Content-Type' => 'text/html; charset=UTF-8'
         );
 
         $buMail->mime->setHTMLBody($message);
 
         $buMail->mime->addAttachment($quoteFile, 'application/pdf');
 
-        $body = $buMail->mime->get();
+        $mime_params = array(
+            'text_encoding' => '7bit',
+            'text_charset' => 'UTF-8',
+            'html_charset' => 'UTF-8',
+            'head_charset' => 'UTF-8'
+        );
+        $body = $buMail->mime->get($mime_params);
 
         $hdrs = $buMail->mime->headers($hdrs);