+++ cnccode/business_classes/BUProblemSLA.inc.php
@@ -416,12 +416,19 @@ class BUProblemSLA extends Business
         $hdrs = array(
             'From' => $senderEmail,
             'Subject' => 'Special Attention Period Ends Soon For ' . $customer->cus_name,
-            'Date' => date("r")
+            'Date' => date("r"),
+            'Content-Type' => 'text/html; charset=UTF-8'
         );
 
         $buMail->mime->setHTMLBody($body);
 
-        $body = $buMail->mime->get();
+        $mime_params = array(
+            'text_encoding' => '7bit',
+            'text_charset' => 'UTF-8',
+            'html_charset' => 'UTF-8',
+            'head_charset' => 'UTF-8'
+        );
+        $body = $buMail->mime->get($mime_params);
 
         $hdrs = $buMail->mime->headers($hdrs);
 
@@ -493,12 +500,19 @@ class BUProblemSLA extends Business
                 'To' => $toEmail,
                 'From' => $senderEmail,
                 'Subject' => 'WARNING - SR for ' . $dbeJProblem->getValue('customerName') . 'assigned to ' . $dbeJProblem->getValue('engineerName') . ' close to breaching SLA',
-                'Date' => date("r")
+                'Date' => date("r"),
+                'Content-Type' => 'text/html; charset=UTF-8'
             );
 
             $buMail->mime->setHTMLBody($body);
 
-            $body = $buMail->mime->get();
+            $mime_params = array(
+                'text_encoding' => '7bit',
+                'text_charset' => 'UTF-8',
+                'html_charset' => 'UTF-8',
+                'head_charset' => 'UTF-8'
+            );
+            $body = $buMail->mime->get($mime_params);
 
             $hdrs = $buMail->mime->headers($hdrs);
 
@@ -627,12 +641,19 @@ class BUProblemSLA extends Business
                 'From' => $senderEmail,
                 'To' => $toEmail,
                 'Subject' => CONFIG_SERVICE_REQUEST_DESC . ' ' . $dbeJCallActivity->getValue('problemID') . ' - Pending Closure on ' . Controller::dateYMDtoDMY($completeDate),
-                'Date' => date("r")
+                'Date' => date("r"),
+                'Content-Type' => 'text/html; charset=UTF-8'
             );
 
         $buMail->mime->setHTMLBody($body);
 
-        $body = $buMail->mime->get();
+        $mime_params = array(
+            'text_encoding' => '7bit',
+            'text_charset' => 'UTF-8',
+            'html_charset' => 'UTF-8',
+            'head_charset' => 'UTF-8'
+        );
+        $body = $buMail->mime->get($mime_params);
 
         $hdrs = $buMail->mime->headers($hdrs);