+++ cnccode/controller_classes/CTActivity.inc.php
@@ -323,6 +323,9 @@ class CTActivity extends CTCNC
             case 'updateHistoricUserTimeLogs':
                 $this->updateHistoricUserTimeLogs();
                 break;
+            case 'test':
+                $this->buActivity->sendSalesRequestAlertEmail(387378, null);
+                break;
 
             case CTCNC_ACT_DISPLAY_SEARCH_FORM:
             default:
@@ -3814,13 +3817,20 @@ class CTActivity extends CTCNC
 
                         }
 
-                        if (
-                            $dsCallActivity->getValue('callActTypeID') == CONFIG_REMOTE_TELEPHONE_ACTIVITY_TYPE_ID &&
-                            $durationHours > $dsHeader->getValue('remoteSupportWarnHours')
-                        ) {
-                            $this->formError = true;
-                            $this->userWarned = true;
-                            $this->dsCallActivity->setMessage('endTime', 'Warning: Activity duration exceeds ' . $dsHeader->getValue('remoteSupportWarnHours') . ' hours');
+                        if ($dsCallActivity->getValue('callActTypeID') == CONFIG_REMOTE_TELEPHONE_ACTIVITY_TYPE_ID) {
+                            if ($durationHours > $dsHeader->getValue('remoteSupportWarnHours')) {
+                                $this->formError = true;
+                                $this->userWarned = true;
+                                $this->dsCallActivity->setMessage('endTime', 'Warning: Activity duration exceeds ' . $dsHeader->getValue('remoteSupportWarnHours') . ' hours');
+                            }
+
+                            $minHours = $dsHeader->getValue(DBEHeader::RemoteSupportMinWarnHours);
+
+                            if ($durationHours < $minHours) {
+                                $this->formError = true;
+                                $this->userWarned = true;
+                                $this->dsCallActivity->setMessage('endTime', 'Remote support under ' . (floor($minHours * 60)) . ' minutes, should this be Customer Contact instead?”.');
+                            }
 
                         }
                     }