+++ cnccode/controller_classes/CTServiceDeskReportCustomer.inc.php
@@ -69,12 +69,19 @@ class CTServiceDeskReportCustomer extends CTCNC
                     'From' => $senderEmail,
                     'To' => $toEmail,
                     'Subject' => 'Service Desk Report ' . $this->buServiceDeskReport->getPeriod() . ' - ' . $this->buServiceDeskReport->getCustomerName(),
-                    'Date' => date("r")
+                    'Date' => date("r"),
+                    'Content-Type' => 'text/html; charset=UTF-8'
                 );
 
                 $buMail->mime->setHTMLBody($report);
 
-                $body = $buMail->mime->get();
+                $mime_params = array(
+                    'text_encoding' => '7bit',
+                    'text_charset' => 'UTF-8',
+                    'html_charset' => 'UTF-8',
+                    'head_charset' => 'UTF-8'
+                );
+                $body = $buMail->mime->get($mime_params);
 
                 $hdrs = $buMail->mime->headers($hdrs);