+++ cnccode/controller_classes/CTServiceRequestsByCustomerReport.inc.php
@@ -191,12 +191,19 @@ class CTServiceRequestsByCustomerReport extends CTCNC
             $hdrs = array(
                 'From' => $senderEmail,
                 'Subject' => $subject,
-                'Date' => date("r")
+                'Date' => date("r"),
+                'Content-Type' => 'text/html; charset=UTF-8'
             );
 
             $buMail->mime->setHTMLBody($body);
 
-            $body = $buMail->mime->get();
+            $mime_params = array(
+                'text_encoding' => '7bit',
+                'text_charset' => 'UTF-8',
+                'html_charset' => 'UTF-8',
+                'head_charset' => 'UTF-8'
+            );
+            $body = $buMail->mime->get($mime_params);
 
             $hdrs = $buMail->mime->headers($hdrs);