+++ cnccode/data_classes/DBEItem.inc.php
@@ -78,7 +78,7 @@ class DBEItem extends DBCNCEntity
             " AND " . $this->getDBColumnName('discontinuedFlag') . " <> 'Y'" .
             " ORDER BY " . $this->getDBColumnName('description') .
             " LIMIT 0,200";
-
+        var_dump($queryString);
         $this->setQueryString($queryString);
 
         $ret = (parent::getRows());
@@ -103,7 +103,6 @@ class DBEItem extends DBCNCEntity
             " FROM " . $this->getTableName() .
             " WHERE " . $this->getDBColumnName('partNo') . " LIKE " . $this->getFormattedLikeValue('partNo') .
             " AND " . $this->getDBColumnName('discontinuedFlag') . " <> 'Y'";
-
         if ($renewalTypeID) {
             $queryString .= " AND renewalTypeID = $renewalTypeID";
         }